move constants out of index
