+++ src/main/java/com/amazon/opendistroforelasticsearch/knn/common/KNNConstants.java
@@ -1,5 +1,5 @@
 /*
- *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *   Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  *   Licensed under the Apache License, Version 2.0 (the "License").
  *   You may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.knn.index.util;
+package com.amazon.opendistroforelasticsearch.knn.common;
 
 public class KNNConstants {
     public static final String SPACE_TYPE = "spaceType";
